Correções e atualizações nos scripts de instalação de programas.
Progs/
- Biblioteca readline adicionada ao script 00_libs.
- Checagem, atualização de versões e correção dos scripts [00 - 05, 09, 11, 12, 13]
- Parametrização de versões e checagem de erros dos scripts [00, 03 - 05, 09, 11]
- Criação de um gitignore e remoção dos Logs do repositório remoto.

+++ Progs/06_python2.7.sh
@@ -3,40 +3,85 @@
 mkdir -p /tmp/python_2
 cd /tmp/python_2
 
-wget https://www.python.org/ftp/python/2.7.11/Python-2.7.11.tgz
+major=2
+minor=7
+micro=16
 
-if [ ! -f Python-2.7.11.tgz ];then
-	echo "Falha ao baixar o arquivo de instalação do Python-2"
+version=$major.$minor.$micro
+pkg_name=Python-$version
+
+echo "Baixando Python $version"
+wget -nv https://www.python.org/ftp/python/$version/$pkg_name.tgz
+
+if [ ! -f $pkg_name.tgz ];then
+	echo "[ERRO] Falha ao baixar o arquivo de instalação do Python $major"
 	exit 1
 fi
+echo "[SUCESSO] Download do pacote $pkg_name efetuado com sucesso"
 
-tar -xf Python-2.7.11.tgz 
-cd Python-2.7.11/
+echo -e "\nDescompactando $pkg_name"
+tar -xf $pkg_name.tgz 
+cd $pkg_name/
 
-./configure --prefix=/opt/python2.7
-make -j8
-make install
+echo -e "\nConfigurando $pkg_name"
+./configure -q --prefix=/opt/python$major.$minor
 
-#install pip for python2.7
-module load python2.7
+if [ $? != 0 ]; then
+  echo "[ERRO] Falha na configuração do Python $major"
+  exit 1
+fi
+echo "[SUCESSO] Python $major configurado com sucesso"
 
-wget https://bootstrap.pypa.io/get-pip.py
+echo -e "\nInstalando Python $major"
+make -s -j8 >/dev/null
+make -s install >/dev/null
+
+if [ $? != 0 ]; then
+        echo "[ERRO] Falha ao instalar o Python $major"
+        exit 1
+fi
+echo "[SUCESSO] Python $major instalado com sucesso"
+
+#install pip for python2
+module load python$major.$minor
+
+echo "Baixando pip"
+wget -nv https://bootstrap.pypa.io/get-pip.py
+
+if [ ! -f get-pip.py ];then
+        echo "[ERRO] Falha ao baixar o arquivo do pip"
+        exit 1
+fi
+echo "[SUCESSO] Download do arquivo do pip efetuado com sucesso"
+
+echo -e "\nInstalando pip e bibliotecas uteis do Python $major"
 python get-pip.py
 
-#installing useful libraries for python 2.7
-pip install numpy 
-pip install pandas
-pip install scipy
-pip install sklearn
-pip install --upgrade https://storage.googleapis.com/tensorflow/linux/gpu/tensorflow-0.8.0rc0-cp27-none-linux_x86_64.whl
-pip install biopython
+for package in numpy pandas scipy sklearn tensorflow biopython
+do
+	pip install -q --upgrade $package
+	if [ $? != 0 ]; then
+		echo "[ERRO] Falha na instalação da biblioteca $package"
+		exit 1
+	fi
+	echo "[SUCESSO] Biblioteca $package instalada com sucesso"
+done
 
 cd ..
-rm Python-2.7.11.tgz
+rm $pkg_name.tgz
 
-rocks create package /opt/python2.7 python2.7
+echo -e "\nCriando pacote"
+rocks create package /opt/python$major.$minor Python release=1 version=$major.$minor.$micro 2>&1 | tail -n 8
 
-$BASE_DIR/AuxScripts/addPackageExtend.sh $(ls python*.rpm)
+if [ ! -f $pkg_name-1.x86_64.rpm ];then
+        echo "[ERRO] Falha ao gerar o pacote rpm do Python $major"
+        exit 1
+fi
+
+$BASE_DIR/AuxScripts/addPackageExtend.sh $(ls Python*.rpm)
+echo "[SUCESSO] Pacote $pkg_name criado e movido com sucesso."
 
 cd /tmp
 rm -rf python_2
+
+exit 0
